load as f64 when using sht utils

+++ src/megatop/pipeline/mask_checker.py
@@ -12,7 +12,7 @@ from megatop.utils import logger, mask, mock
 def spin_derivatives(manager: DataManager):
     common_hnits_map = hp.read_map(manager.path_to_common_nhits_map)
     binary_mask = hp.read_map(manager.path_to_binary_mask)
-    analysis_mask = hp.read_map(manager.path_to_analysis_mask)
+    analysis_mask = hp.read_map(manager.path_to_analysis_mask, dtype=np.float64)
 
     # Spin derivatives computation
     first_custom, second_custom = mask.get_spin_derivatives(analysis_mask)
@@ -44,7 +44,7 @@ def generate_mask_sim(manager: DataManager, config: Config, int_n_sim):
     rank = comm.Get_rank()
 
     Cl_cmb_model = mock.get_Cl_CMB_model_from_manager(manager)
-    analysis_mask = hp.read_map(manager.path_to_analysis_mask)
+    analysis_mask = hp.read_map(manager.path_to_analysis_mask, dtype=np.float64)
 
     Cl_cmb_model_pureB = Cl_cmb_model.copy()
     Cl_cmb_model_pureB[[1]] = 0.0
